Make my first commit message
The reason for this commit message is to solve two things:

    - Add additional details to the README.md file in the project
    - Get in the habit of making a good commit message

This message is coming from a result of learning how to make a good
commit message that is helpful to others when collaborating on a
project.

I've followed a template provided by one of the external links from
TOP and am attempting to explain what I am changing and why.

Cheers.