Add multiple elements to snickerdoodle.html
Having followed directions from TOP exercise, I added a description
under the heading under the picutre, an unordered list detailing
the ingredients, and an ordered list for the instructions. I also
made sure to add a convenient link to go back to the home page, as well
as a link that takes the user to the original recipe page.